+++ models/experimental/deit/deit_cpp/helper_funcs.cpp
@@ -368,4 +368,26 @@ torch::Tensor load_and_preprocess_image(
     return tensor;
 }
 
+/**
+ * Compute Pearson Correlation Coefficient (PCC) between two tensors
+ * @param tensor1 First tensor
+ * @param tensor2 Second tensor
+ * @return PCC value
+ */
+double compute_pcc(const torch::Tensor& tensor1, const torch::Tensor& tensor2) {
+    auto flat1 = tensor1.flatten().to(torch::kFloat32);
+    auto flat2 = tensor2.flatten().to(torch::kFloat32);
+    
+    auto mean1 = flat1.mean();
+    auto mean2 = flat2.mean();
+    
+    auto centered1 = flat1 - mean1;
+    auto centered2 = flat2 - mean2;
+    
+    auto numerator = (centered1 * centered2).sum();
+    auto denominator = torch::sqrt((centered1 * centered1).sum() * (centered2 * centered2).sum());
+    
+    return numerator.item<double>() / denominator.item<double>();
+}
+
 } // namespace helper_funcs
