+++ models/experimental/deit/deit_cpp/test_cpp/test_deit_self_output.cpp
@@ -15,6 +15,7 @@
 #include <ttnn/tensor/tensor.hpp>
 #include <ttnn/types.hpp>
 #include <ttnn/device.hpp>
+#include <ttnn/distributed/api.hpp>
 #include <tt-metalium/host_api.hpp>
 #include <tt-metalium/device.hpp>
 
@@ -23,29 +24,6 @@
 #include "../helper_funcs.h"
 
 namespace {
-
-/**
- * Compute Pearson Correlation Coefficient (PCC) between two tensors
- * @param tensor1 First tensor
- * @param tensor2 Second tensor
- * @return PCC value
- */
-double compute_pcc(const torch::Tensor& tensor1, const torch::Tensor& tensor2) {
-    auto flat1 = tensor1.flatten().to(torch::kFloat32);
-    auto flat2 = tensor2.flatten().to(torch::kFloat32);
-    
-    auto mean1 = flat1.mean();
-    auto mean2 = flat2.mean();
-    
-    auto centered1 = flat1 - mean1;
-    auto centered2 = flat2 - mean2;
-    
-    auto numerator = (centered1 * centered2).sum();
-    auto denominator = torch::sqrt((centered1 * centered1).sum() * (centered2 * centered2).sum());
-    
-    return numerator.item<double>() / denominator.item<double>();
-}
-
 
 /**
  * Test DeiT Self Output inference
@@ -116,7 +94,6 @@ void test_deit_self_output_inference(const std::string& model_path) {
     // DeiT uses 196 patches + 1 CLS token + 1 distillation token = 198 tokens
     torch::Tensor hidden_states = torch::randn({1, 198, 768}, torch::kFloat32);
 
-    
     // Call self-output module forward
     // The PyTorch forward method only takes hidden_states (self is implicit)
     std::vector<torch::jit::IValue> inputs;
@@ -133,7 +110,6 @@ void test_deit_self_output_inference(const std::string& model_path) {
     // Convert inputs to TT tensors
     auto tt_hidden_states = helper_funcs::torch_to_tt_tensor_tile(hidden_states, device);
 
-    
     // Run TT model inference
     auto tt_out = tt_self_output.forward(tt_hidden_states);
     
@@ -144,7 +120,7 @@ void test_deit_self_output_inference(const std::string& model_path) {
     tt_output_torch = tt_output_torch.squeeze(0); // Remove batch dimension
     
     // Compute PCC between PyTorch and TT outputs
-    double pcc = compute_pcc(torch_output, tt_output_torch);
+    double pcc = helper_funcs::compute_pcc(torch_output, tt_output_torch);
     
     // Check if PCC meets threshold
     if (pcc >= pcc_threshold) {
@@ -153,6 +129,8 @@ void test_deit_self_output_inference(const std::string& model_path) {
         std::cout << "Test FAILED! PCC: " << pcc << " (threshold: " << pcc_threshold << ")" << std::endl;
     }
 
+    // Clean up device resources
+    ttnn::distributed::close_mesh_device(device);
 }
 
 } // anonymous namespace
